complete the style for (C) checkbox
