+++ src/components/Checkbox/interface.ts
@@ -1,6 +1,8 @@
 import { InputProps } from "@/common";
 
 export interface CheckboxProps extends Omit<InputProps, "type" | "ref"> {
-  classNameContainer?: string;
-  label?: string;
+	classNameContainer?: string;
+	label?: string;
+	indeterminate?: boolean;
+	position?: "left" | "right";
 }
